feat: :sparkles: UserProfile 자잘한 styling (capitalize, coalition mark 등)
